osd fs: __r4w_get_page rely on PageUptodate for uptodate
The forthcoming patch "mm: migrate dirty page without
clear_page_dirty_for_io etc" simplifies the migration of a PageDirty
pagecache page: one stat needs moving from zone to zone and that's about
all.

It's convenient and safest for it to shift the PageDirty bit from old page
to new, just before updating the zone stats: before copying data and
marking the new PageUptodate.  This is all done while both pages are
isolated and locked, just as before; and just as before, there's a moment
when the new page is visible in the radix_tree, but not yet PageUptodate. 
What's new is that it may now be briefly visible as PageDirty before it is
PageUptodate.

When I scoured the tree to see if this could cause a problem anywhere, the
only places I found were in two similar functions __r4w_get_page(): which
look up a page with find_get_page() (not using page lock), then claim it's
uptodate if it's PageDirty or PageWriteback or PageUptodate.

I'm not sure whether that was right before, but now it might be wrong (on
rare occasions): only claim the page is uptodate if PageUptodate.  Or
perhaps the page in question could never be migratable anyway?

Signed-off-by: Hugh Dickins <hughd@google.com>
Cc: Boaz Harrosh <ooo@electrozaur.com>
Cc: Boaz Harrosh <boaz@plexistor.com>
Cc: Benny Halevy <bhalevy@panasas.com>
Cc: Trond Myklebust <trond.myklebust@primarydata.com>
Cc: Christoph Lameter <cl@linux.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>