lib/vsprintf.c: warn about too large precisions and field widths
The field width is overloaded to pass some extra information for some %p
extensions (e.g.  #bits for %pb).  But we might silently truncate the
passed value when we stash it in struct printf_spec (see e.g. 
"lib/vsprintf.c: expand field_width to 24 bits").  Hopefully 23 value bits
should now be enough for everybody, but if not, let's make some noise.

Do the same for the precision.  In both cases, clamping seems more
sensible than truncating.  A negative precision is the same as 0, so use
that as lower bound.  For the field width, the smallest representable
value is actually -(1<<23), but a negative field width means 'set the LEFT
flag and use the absolute value', so we want the absolute value to fit.

Signed-off-by: Rasmus Villemoes <linux@rasmusvillemoes.dk>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>