fs: clear file privilege bits when mmap writing
Normally, when a user can modify a file that has setuid or setgid bits,
those bits are cleared when they are not the file owner or a member of the
group.  This is enforced when using write and truncate but not when
writing to a shared mmap on the file.  This could allow the file writer to
gain privileges by changing a binary without losing the setuid/setgid/caps
bits.

Changing the bits requires holding inode->i_mutex, so it cannot be done
during the page fault (due to mmap_sem being held during the fault). 
Instead, clear the bits if PROT_WRITE is being used at mmap open time. 
But we can't do the check in the right place inside mmap, so we have to do
it before holding mmap_sem, which means duplicating some checks, which
have to be available to the non-MMU builds too.

Signed-off-by: Kees Cook <keescook@chromium.org>
Cc: Jan Kara <jack@suse.cz>
Cc: Willy Tarreau <w@1wt.eu>
Cc: "Eric W. Biederman" <ebiederm@xmission.com>
Cc: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Cc: Oleg Nesterov <oleg@redhat.com>
Cc: Rik van Riel <riel@redhat.com>
Cc: Chen Gang <gang.chen.5i5j@gmail.com>
Cc: Davidlohr Bueso <dave@stgolabs.net>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: Alexander Viro <viro@ftp.linux.org.uk>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>